solutions/romance-unzipped/pipeline/requirements.txt
anthropic
openai
Pillow
PyYAML
pytest
python-dotenv
mlx-whisper
yt-dlp
